ChatBot/requirements.txt
fastapi
uvicorn
torch>=2.0.0
transformers
jinja2
accelerate
bitsandbytes
huggingface_hub
gradio>=4.0.0
llama-cpp-python
huggingface_hub
pandas
